desafio03.py

+++ desafio03.py
@@ -1,8 +1,8 @@
 print("Vamos montar uma soma?")
-primeiro_numero = int(input("Digite o primeiro número: "))
-segundo_numero = int(input("Digite o segundo número: "))
+primeiro_numero = float(input("Digite o primeiro número: "))
+segundo_numero = float(input("Digite o segundo número: "))
 
-resultado = int(primeiro_numero+segundo_numero)
+resultado = float(primeiro_numero+segundo_numero)
 print(f"Restudado: {resultado}")
